requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
botorch
fire
gpytorch
matplotlib
numpy
pandas
pip
torch
pyyaml
scikit-learn
scipy
